Added feature to take only backup

+++ powershell_scripts/db_cloning_automation/clone_db.ps1
@@ -1,9 +1,11 @@
 ﻿
 param(
 [Parameter(Mandatory=$True)]
+[string]$ddb,
+[Parameter(Mandatory=$False)]
 [string]$sdb,
-[Parameter(Mandatory=$True)]
-[string]$ddb
+[Parameter(Mandatory=$False)]
+[switch]$backup
 )
 
 # Getting database information from json file
@@ -12,7 +14,7 @@ $dbFilePath = $curDir + "\andatabases.json"
 $allDbsJson = ConvertFrom-Json "$(get-content $dbFilePath)"
 
 # Login to Azure Account
-Login-AzureRmAccount
+#Login-AzureRmAccount
 Select-AzureRmSubscription -SubscriptionID $allDbsJson.subscriptionId
 
 # Storage account info to store BACPAC
@@ -39,26 +41,30 @@ while ($exportStatus.Status -ne 'Succeeded')
 
 Write-Host "Backup completed, URL:" $BacpacUri -ForegroundColor Green
 
-Write-Host "Deleting database" $allDbsJson.Databases.$ddb.dbname -ForegroundColor Yellow
 
-Remove-AzureRmSqlDatabase -ResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup -ServerName $allDbsJson.Databases.$ddb.server `
-    -DatabaseName $allDbsJson.Databases.$ddb.dbname -Force
 
-Write-Host "Copying database to destination" -ForegroundColor Yellow
-
-# Copy source database to the target server
-$databasecopy = New-AzureRmSqlDatabaseCopy -ResourceGroupName $allDbsJson.Databases.$sdb.resourcegroup -ServerName $allDbsJson.Databases.$sdb.server `
-                    -DatabaseName $allDbsJson.Databases.$sdb.dbname -CopyResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup `
-                    -CopyServerName $allDbsJson.Databases.$ddb.server -CopyDatabaseName $allDbsJson.Databases.$ddb.dbname `
-                    -ServiceObjectiveName "S0"
-
-# Adding to the respective elastic pool
-Write-Host "Adding database to Elastic Pool if any" -ForegroundColor Yellow
-if ($allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).ElasticPoolName -ne "NA")
+if(-not($backup))
 {
-    Set-AzureRmSqlDatabase -ResourceGroupName "AN-PREPROD" `
-    -ServerName $allDbsJson.Databases.$ddb.server `
-    -DatabaseName $allDbsJson.Databases.$ddb.dbname `
-    -ElasticPoolName $allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).ElasticPoolName
-}
+    Write-Host "Deleting database" $allDbsJson.Databases.$ddb.dbname -ForegroundColor Yellow
+    Remove-AzureRmSqlDatabase -ResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup -ServerName $allDbsJson.Databases.$ddb.server `
+        -DatabaseName $allDbsJson.Databases.$ddb.dbname -Force
 
+    Write-Host "Copying database to destination" -ForegroundColor Yellow
+
+    # Copy source database to the target server
+    $databasecopy = New-AzureRmSqlDatabaseCopy -ResourceGroupName $allDbsJson.Databases.$sdb.resourcegroup -ServerName $allDbsJson.Databases.$sdb.server `
+                        -DatabaseName $allDbsJson.Databases.$sdb.dbname -CopyResourceGroupName $allDbsJson.Databases.$ddb.resourcegroup `
+                        -CopyServerName $allDbsJson.Databases.$ddb.server -CopyDatabaseName $allDbsJson.Databases.$ddb.dbname `
+                        -ServiceObjectiveName "S0"
+
+    # Adding to the respective elastic pool
+    Write-Host "Adding database to Elastic Pool if any" -ForegroundColor Yellow
+    if ($allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).ElasticPoolName -ne "NA")
+    {
+        Set-AzureRmSqlDatabase -ResourceGroupName "AN-PREPROD" `
+        -ServerName $allDbsJson.Databases.$ddb.server `
+        -DatabaseName $allDbsJson.Databases.$ddb.dbname `
+        -ElasticPoolName $allDbsJson.Credentials.($allDbsJson.Databases.$ddb.server).ElasticPoolName
+    }
+
+}
